you can block and unblock users, update password

+++ src/Controllers/UserController.php
@@ -2,14 +2,14 @@
 
 namespace Gmlo\CMS\Controllers;
 
+use Carbon\Carbon;
 use Gmlo\CMS\Modules\Users\UsersRepo;
 use Gmlo\CMS\Requests\CreateUserRequest;
+use Gmlo\CMS\Requests\UpdatePassRequest;
 use Gmlo\CMS\Requests\UpdateUserRequest;
-use Illuminate\Http\Request;
-
+use Illuminate\Contracts\Auth\Guard;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
-use Gmlo\CMS\Modules\Users\User;
 
 
 class UserController extends Controller
@@ -100,11 +100,55 @@ class UserController extends Controller
      * @param  int  $id
      * @return Response
      */
-    public function destroy($id)
+    public function destroy(Guard $guard, $id)
     {
         $user = $this->usersRepo->findOrFail($id);
+        if($user->id == $guard->user()->id)
+        {
+            \Alert::message("You can't delete yourself!", "danger");
+            return redirect()->back();
+        }
         $this->usersRepo->delete($user);
         \Alert::message("User deleted!");
         return redirect()->route('CMS::admin.users.index');
+    }
+
+    public function editPassword($id)
+    {
+        $user = $this->usersRepo->findOrFail($id);
+
+        return view('CMS::users.edit-password', compact('user'));
+    }
+
+    public function updatePassword(UpdatePassRequest $request, $id)
+    {
+        $user = $this->usersRepo->findOrFail($id);
+        $user->password = \Hash::make($request->get('password'));
+        $this->usersRepo->save($user);
+        \Alert::message("Password updated!");
+        return redirect()->route('CMS::admin.users.edit', $user->id);
+    }
+
+    public function statusToggle(Guard $guard, $id)
+    {
+        $user = $this->usersRepo->findOrFail($id);
+        if($user->isBlocked())
+        {
+            $user->blocked_at = null;
+            \Alert::message("User unblocked!");
+        }
+        else
+        {
+            if($user->id == $guard->user()->id)
+            {
+                \Alert::message("You can't block yourself!", "danger");
+                return redirect()->back();
+            }
+            $user->blocked_at = Carbon::now();
+            \Alert::message("User blocked!");
+        }
+        $this->usersRepo->save($user);
+
+        return redirect()->route('CMS::admin.users.edit', $user->id);
     }
 }
